+++ CHANGELOG.md
@@ -44,6 +44,7 @@ This release has an [MSRV] of 1.82.
 
 - Fix text editing for layouts which contain inline boxes ([#299][] by [@valadaptive][])
 - Fix cursor navigation in RTL text sometimes getting stuck within a line ([#331][] by [@valadaptive][])
+- Using `Layout::align` on an aligned layout without breaking lines again. ([#342][] by [@xStrom][])
 
 ## [0.3.0] - 2025-02-27
 
@@ -173,6 +174,7 @@ This release has an [MSRV] of 1.70.
 [@waywardmonkeys]: https://github.com/waywardmonkeys
 [@wfdewith]: https://github.com/wfdewith
 [@xorgy]: https://github.com/xorgy
+[@xStrom]: https://github.com/xStrom
 
 [#54]: https://github.com/linebender/parley/pull/54
 [#55]: https://github.com/linebender/parley/pull/55
@@ -232,6 +234,7 @@ This release has an [MSRV] of 1.70.
 [#315]: https://github.com/linebender/parley/pull/315
 [#318]: https://github.com/linebender/parley/pull/318
 [#331]: https://github.com/linebender/parley/pull/331
+[#342]: https://github.com/linebender/parley/pull/342
 
 [Unreleased]: https://github.com/linebender/parley/compare/v0.3.0...HEAD
 [0.3.0]: https://github.com/linebender/parley/releases/tag/v0.3.0